refactor(test): move usage and insights helpers to db-test-seeders and db-test-assertions
Move 12 remaining usage/insights helpers from api-test-helpers/credits.ts
to their proper locations:
- 7 seeders to db-test-seeders/credits.ts (with @why-db-direct annotations)
- 5 assertion helpers to db-test-assertions/credits.ts

After this change, api-test-helpers/credits.ts contains only service
wrappers and re-exports with zero db/schema imports.

Part 2 of #9348 (parent: #9321)
